Explanation of changes:
We use a list comprehension [tag.text for tag in quote.find_all("a", class_="tag")] to concisely extract the tag names. This is a more Pythonic way to create a list.
We store the quotes, authors, and tags in a list of dictionaries (all_quotes) to keep the data organized.
We use all_quotes.extend(...) to add the new quotes from each page to the main list.
The printing part now includes the tags for each quote.
This enhanced version not only extracts more data but also demonstrates more efficient and readable Python code. You can further adapt this code to store the data in a file or database for later analysis or use.

Feel free to suggest any other modifications or features you'd like to add!